REFACTOR: Remove bitdubai from middleware asset factory plugin. #5800
