--- DAP/plugin/middleware/fermat-dap-plugin-middleware-asset-factory-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/middleware/asset/issuer/developer/bitdubai/version_1/exceptions/MissingAssetDataException.java
@@ -1,14 +1,0 @@
-package com.bitdubai.fermat_dap_plugin.layer.middleware.asset.issuer.developer.bitdubai.version_1.exceptions;
-
-import com.bitdubai.fermat_api.FermatException;
-
-/**
- * Created by franklin on 15/09/15.
- */
-public class MissingAssetDataException extends FermatException{
-    public static final String DEFAULT_MESSAGE = "There was an error trying to insert data in table. The referenced Asset Facory public key does not exists.";
-
-    public MissingAssetDataException(String message, Exception cause, String context, String possibleReason) {
-        super(message, cause, context, possibleReason);
-    }
-}
