+++ DAP/plugin/middleware/fermat-dap-plugin-middleware-asset-factory/build.gradle
